requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
boto3==1.26.130
botocore==1.29.130
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.3
dataclass-wizard==0.22.2
ecdsa==0.18.0
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
Mako==1.2.4
mangum==0.17.0
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pg8000==1.29.8
platformdirs==3.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
scramp==1.4.4
sentry-sdk==1.26.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
sqlalchemy-easy-softdelete==0.8.2
starlette==0.26.1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
yarl==1.9.2
